Improvments
Locked the slider value for development.
Changed the code style to Google Java code style.
Context stoled in SysUICode.
Added ViewGroupUtils.
Added ColorPicker layout.
Changed the Material design lib to rey's material lib.

+++ app/src/main/java/com/rohden/recentsapp/FrameResCode.java
@@ -16,8 +16,8 @@ public class FrameResCode implements IXposedHookZygoteInit{
         //-----------Shared preferences------------//
 		XSharedPreferences pref = new XSharedPreferences("com.example.recentsapp", "user_settings");
         
-        int width=pref.getInt("recent_width",0);
-        int height=pref.getInt("recent_height",0);
+        float width=pref.getFloat("recent_width",0);
+        float height=pref.getFloat("recent_height",0);
         //-----------------------------------------//
         
         //these values are from systemUI recent apps thumbnails, and seems to be on framework-res.apk , why?i don't know.
